pyproject.toml
[project]
name = "kube-autotuner"
version = "0.1.0"
description = "Benchmark-driven kernel parameter tuning for Kubernetes nodes"
readme = "README.md"
requires-python = ">=3.14"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    { name = "Jacob Colvin", email = "jacobcolvin1@gmail.com" },
]
dependencies = [
    "kubernetes>=31.0.0",
    "pydantic>=2.12",
    "pyyaml>=6.0",
    "typer>=0.23",
]

[project.scripts]
kube-autotuner = "kube_autotuner.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.0",
    "pytest-timeout>=2.4",
    "types-PyYAML>=6.0",
]
optimize = ["ax-platform>=1.1"]
analysis = [
    "pandas>=3.0",
    "plotly>=6.0",
    "scikit-learn>=1.7",
]

[build-system]
requires = ["uv_build>=0.11,<0.12"]
build-backend = "uv_build"

[tool.uv]
default-groups = "all"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py314"
extend-exclude = [".task"]
preview = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "ISC001",
    "D203",
    "D213",
    "CPY001",
    "TD002",
    "TD003",
    "FIX002",
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",
    "S404",
    "D1",
    "DOC501",
    "EM101",
    "EM102",
    "PLC1901",
    "PLR2004",
    "PLR6301",
    "SLF001",
    "ANN",
    "INP001",
]
# Integration tests drive subprocess calls and interact with mutable global
# state (env vars, kubeconfig). The subprocess S603/S607 and docstring
# return/yield rules (DOC201/DOC402) are noise in this context; blanket-
# ignoring them here keeps per-call ``# noqa`` off the hot paths.
"tests/integration/**/*.py" = [
    "S603",
    "S607",
    "DOC201",
    "DOC402",
]
"**/__init__.py" = [
    "D104",
    "F401",
]
"**/cli.py" = [
    "FBT",
    "PLR0913",
    "PLR0915",
    "PLR0917",
]
# K8sClient methods route API errors through a NoReturn ``_raise`` helper,
# which DOC502 cannot trace. Documenting the real raise paths on the public
# surface is more useful than hiding them.
"src/kube_autotuner/k8s/client.py" = [
    "DOC502",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-type-checking]
strict = true
runtime-evaluated-base-classes = ["pydantic.BaseModel"]

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = false
allow-star-arg-any = false
mypy-init-return = false

[tool.ruff.lint.pylint]
max-args = 6
max-branches = 12
max-returns = 6
max-statements = 40

[tool.ruff.lint.isort]
known-first-party = ["kube_autotuner"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
minversion = "9.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=kube_autotuner",
    "--cov-report=term-missing",
    "-m",
    "not integration",
]
markers = [
    "integration: requires a Talos Docker cluster",
    "requires_real_sysctl_write: requires a cluster where privileged pods can write host sysctls",
]
timeout = 120
testpaths = ["tests"]

[tool.ty.src]
include = ["src", "tests"]

[tool.ty.rules]
all = "error"

[tool.ty.environment]
python-version = "3.14"
python-platform = "linux"
root = ["src"]

[[tool.ty.overrides]]
include = ["tests/**"]

[tool.ty.overrides.rules]
possibly-unresolved-reference = "warn"
unused-ignore-comment = "warn"

# Ax lives in the `optimize` group. It is pulled by default via
# `tool.uv.default-groups = "all"`, but CI jobs and wheel consumers may skip
# the group with `--no-group optimize`; the lazy-import pattern plus this
# override keep type-checking green in that case.
[[tool.ty.overrides]]
include = ["src/kube_autotuner/optimizer.py"]

[tool.ty.overrides.rules]
unresolved-import = "ignore"

# pandas / plotly / scikit-learn live in the `analysis` group. Same deal as
# `optimizer.py`: installed by default, but opt-out with `--no-group analysis`
# must still type-check thanks to the lazy-import discipline.
[[tool.ty.overrides]]
include = [
    "src/kube_autotuner/analysis.py",
    "src/kube_autotuner/report.py",
    "src/kube_autotuner/plots.py",
]

[tool.ty.overrides.rules]
unresolved-import = "ignore"

# python-kubernetes exposes swagger-generated models whose attributes ty
# cannot always resolve statically. Callers outside this module use only
# the fully typed K8sClient / K8sApiError surface, so the override is
# narrow.
[[tool.ty.overrides]]
include = ["src/kube_autotuner/k8s/client.py"]

[tool.ty.overrides.rules]
unresolved-attribute = "warn"
possibly-missing-attribute = "warn"
invalid-assignment = "warn"

# K8sClient tests bypass ``__init__`` and inject MagicMock handles on the
# typed attributes. ty's inferred types come from the real assignments in
# ``__init__``, so assigning canned return values or accessing
# ``assert_called_*`` trips on the typed surface even though the runtime
# just sees ``mock.foo``.
[[tool.ty.overrides]]
include = ["tests/test_k8s_client.py"]

[tool.ty.overrides.rules]
invalid-assignment = "warn"
unresolved-attribute = "warn"
